refactor useRedirectAuthenticatedUsers hook
