+++ features/auth/signup/Signup.tsx
@@ -8,12 +8,11 @@ import MobileAuthBg from "features/auth/resources/mobile-auth-bg.jpg";
 import CommunityGuidelinesForm from "features/auth/signup/CommunityGuidelinesForm";
 import { Trans, useTranslation } from "i18n";
 import { AUTH, GLOBAL } from "i18n/namespaces";
-import { useRouter } from "next/router";
 import { useEffect } from "react";
 import vercelLogo from "resources/vercel.svg";
-import { dashboardRoute, loginRoute, tosRoute } from "routes";
+import { loginRoute, tosRoute } from "routes";
+import { useRedirectAuthenticatedUsers } from "utils/hooks";
 import makeStyles from "utils/makeStyles";
-import stringOrFirstString from "utils/stringOrFirstString";
 
 import { useAuthContext } from "../AuthProvider";
 import useAuthStyles from "../useAuthStyles";
@@ -142,11 +141,10 @@ function CurrentForm() {
 }
 
 export default function Signup() {
+  useRedirectAuthenticatedUsers()
   const { t } = useTranslation([AUTH, GLOBAL]);
-  const router = useRouter();
   const { authState, authActions } = useAuthContext();
   const isSignupComplete = authState.flowState?.isCompleted;
-  const authenticated = authState.authenticated
   const error = authState.error;
   const authClasses = useAuthStyles();
   const classes = useStyles();
@@ -159,19 +157,9 @@ export default function Signup() {
     if (authState.error) window.scroll({ top: 0, behavior: "smooth" });
   }, [authState.error]);
 
-
-  const redirectFrom = stringOrFirstString(router.query.from) ?? dashboardRoute;
-  const redirectTo = isSignupComplete
-    ? loginRoute
-    : authenticated
-      ? redirectFrom === "/" || redirectFrom === "%2F"
-        ? dashboardRoute
-        : redirectFrom
-      : "";
-
   return (
     <>
-      {redirectTo && <Redirect to={redirectTo} />}
+      {isSignupComplete && <Redirect to={loginRoute} />}
       <HtmlMeta title={t("global:sign_up")} />
       <div
         className={classNames(
